refactor: OAuth2 관련 클래스 패키지 이동 ### OAuth2 관련 클래스 이동 - common 패키지로 이동한 클래스: CustomOAuth2User, CustomOAuthUserService, GoogleOAuthResponse, NaverOAuthResponse, OAuth2Response, OAuthLoginController
### 패키지 이동에 따른 import 문 수정
- 영향을 받은 컨트롤러 및 클래스:
Authenticator, CoinChargeHistoryController, CoinUseHistoryController, CommentController, EpisodeController, FavoriteNovelController, MemberController, NovelController, ReCommentController, SettlementController, SseController, 등
