+++ src/main/java/com/ham/netnovel/common/OAuth/OAuthLoginController.java
@@ -1,4 +1,4 @@
-package com.ham.netnovel.OAuth;
+package com.ham.netnovel.common.OAuth;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
