+++ src/main/java/com/ham/netnovel/member/MemberController.java
@@ -1,7 +1,7 @@
 package com.ham.netnovel.member;
 
 
-import com.ham.netnovel.OAuth.CustomOAuth2User;
+import com.ham.netnovel.common.OAuth.CustomOAuth2User;
 import com.ham.netnovel.common.utils.PageableUtil;
 import com.ham.netnovel.member.dto.*;
 import com.ham.netnovel.member.service.MemberMyPageService;
